requirements.txt
Django==4.2.1
psycopg2==2.9.6
djangorestframework==3.14.0
djoser==2.2.0
pillow==9.5.0